#911 update telegram-decode-api and telegram-decode-imp to support telegram version 8.4; Update the telegram parser to support new MSG types and to use long instead of int for mediasize, PhotoSize and remetenteId;
Conflicts:
	iped-parsers/iped-parsers-impl/pom.xml

+++ iped-parsers/iped-parsers-impl/src/main/java/dpf/ap/gpinf/telegramextractor/Extractor.java
@@ -369,7 +369,7 @@ public class Extractor {
         return msgs;
     }
 
-    private void loadDocument(Message message, List<String> names, int size) {
+    private void loadDocument(Message message, List<String> names, long size) {
         for (String name : names) {
             String query = getQuery(name, size);
             IItemBase item = getFileFromQuery(query);
@@ -427,7 +427,7 @@ public class Extractor {
         }
     }
 
-    private String getQuery(String name, int size) {
+    private String getQuery(String name, long size) {
         String query = BasicProps.NAME + ":\"" + searcher.escapeQuery(name) + "\"";
         query += size > 0 ? " && " + BasicProps.LENGTH + ":" + size : "";
         return query;
